Pre-filter candidate compare loadouts by class
