+++ src/app/loadout-builder/generated-sets/GeneratedSetButtons.tsx
@@ -1,18 +1,15 @@
-import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 import { t } from 'app/i18next-t';
 import { DimItem, PluggableInventoryItemDefinition } from 'app/inventory/item-types';
 import { DimStore } from 'app/inventory/store-types';
 import { applyLoadout } from 'app/loadout-drawer/loadout-apply';
 import { editLoadout } from 'app/loadout-drawer/loadout-events';
 import { Loadout } from 'app/loadout-drawer/loadout-types';
-import { convertToLoadoutItem } from 'app/loadout-drawer/loadout-utils';
 import { useD2Definitions } from 'app/manifest/selectors';
 import { useThunkDispatch } from 'app/store/thunk-dispatch';
-import _ from 'lodash';
 import { Dispatch } from 'react';
 import { LoadoutBuilderAction } from '../loadout-builder-reducer';
-import { ArmorSet, LockableBucketHashes } from '../types';
-import { statTier } from '../utils';
+import { ArmorSet } from '../types';
+import { updateLoadoutWithArmorSet } from '../updated-loadout';
 import styles from './GeneratedSetButtons.m.scss';
 
 /**
@@ -42,7 +39,7 @@ export default function GeneratedSetButtons({
 }) {
   const defs = useD2Definitions()!;
   const dispatch = useThunkDispatch();
-  const loadout = () => createLoadout(defs, originalLoadout, set, items, lockedMods);
+  const loadout = () => updateLoadoutWithArmorSet(defs, originalLoadout, set, items, lockedMods);
 
   // Opens the loadout menu for the generated set
   const openLoadout = () =>
@@ -95,41 +92,4 @@ export default function GeneratedSetButtons({
       )}
     </div>
   );
-}
-
-/**
- * Create a new loadout from the original prototype loadout, but with the armor items replaced with this loadout's armor.
- * Used for equipping or creating a new saved loadout.
- */
-function createLoadout(
-  defs: D2ManifestDefinitions,
-  originalLoadout: Loadout,
-  set: ArmorSet,
-  items: DimItem[],
-  lockedMods: PluggableInventoryItemDefinition[]
-): Loadout {
-  const data = {
-    tier: _.sumBy(Object.values(set.stats), statTier),
-  };
-  const existingItemsWithoutArmor = originalLoadout.items.filter(
-    (li) =>
-      !LockableBucketHashes.includes(
-        defs.InventoryItem.get(li.hash)?.inventory?.bucketTypeHash ?? 0
-      )
-  );
-  const loadoutItems = items.map((item) => convertToLoadoutItem(item, true));
-
-  // We need to add in this set's specific stat mods (artifice, general) to the list of user-chosen mods
-  // We can't start with the list of mods in the existing loadout parameters because lockedMods has filtered out
-  // invalid mods, mods that don't fit, and general mods if we're auto-assigning general mods.
-  const allMods = [...lockedMods.map((m) => m.hash), ...set.statMods];
-  return {
-    ...originalLoadout,
-    parameters: {
-      ...originalLoadout.parameters,
-      mods: allMods.length ? allMods : undefined,
-    },
-    items: [...existingItemsWithoutArmor, ...loadoutItems],
-    name: originalLoadout.name ?? t('Loadouts.Generated', data),
-  };
 }
